% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{doSomething}
\alias{doSomething}
\title{Sample skeleton for custom science - Main application function}
\usage{
doSomething(datadir)
}
\arguments{
\item{datadir}{Path to data directory.}
}
\description{
Sample skeleton for custom science - Main application function
}
