% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\name{doSomething}
\alias{doSomething}
\title{Main application entrypoint}
\usage{
doSomething(datadir)
}
\arguments{
\item{datadir}{Path to data directory.}
}
\description{
Main application entrypoint
}

